applied pr suggestions: moved geo types to separate files and scoped under geo namespace.
