+++ Amplify/Categories/Geo/Types/Geo.swift
@@ -1,0 +1,11 @@
+//
+// Copyright Amazon.com Inc. or its affiliates.
+// All Rights Reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+import Foundation
+
+/// Namespace for GeoCategory Types
+public struct Geo {}
